sistema_bar9/requirements.txt
mysql-connector-python==8.0.32
pywin32==306
Pillow==10.0.0
pygame==2.5.0